test: etherscan tests passing with pro or on-pro API key
