+++ pycaw/etherscan/__init__.py
@@ -1,10 +1,8 @@
 """TODO module docs for pycaw.etherscan"""
 from pycaw.etherscan import types 
 from pycaw.etherscan import etherscan_connector
-from pycaw.etherscan import token_info_connector 
 
 EtherscanConnector = etherscan_connector.EtherscanConnector 
-TokenInfoConnector = token_info_connector.TokenInfoConnector
 # TokenInfoConnector.__doc__ = 
 """TODO doc"""
